fix #0: Add new data and update context in TelemetryRollupService.js
